requirements.txt
fastapi
uvicorn[standard]
numpy
faiss-cpu
pypdf
sentence-transformers
google-genai
python-dotenv